UI improvements
Set the entire page hard refresh from 20 mins to 8 hours. No real need for it, but its nice to refresh the html/js/css from time to time in case there was an update on the filesystem.

Made the UI a bit more clear, now it shows MAX and TARGET on the UI to explain what the high range temps actually mean.