Added Bambu API login grabbing useful values
With the Bambu API we can get the following:
- Plate Image
- Model weight
- Total prints

All values have been added to the dashboard UI.

With this update I also added many requirements: cors / express and node-fetch packages are now required.

Your Bambu username and password are optional, if not provided the image will not show of the plate and the values for weight and total prints will remain blank.