Slight adjustment to printer info widget
The "Total print count" on the printer-info widget was inaccurate.

The number is calculated by adding up the history of printed items from the Bambu cloud print history list. The issue with this method of calculation is that Bambu only keeps a running 90 day history of cloud prints.

Instead of removing the feature, I've modified the text to reflect that the total print count is only for the past 90 days of print history.