v1.1.4 - Fix bug with login verification
Bambu is now requiring extra verification on logins. When this occurs the BambuBoard cannot login and retrieve print information (model name, weight, model image). In this version we provide a website that will allow you to enter your username and password, send code, and a place for you to enter the code and login. This will cache your access code so we can use it for future logins. The access code should work for 1 year from what we understand.

This is just a temporary fix until something better can be put in place. To login go to this website:
http://<bambu_printer_ip>/login.html

After logging and and verification of code, please restart BambuBoard to get it working again.